Enc common: separation of warped motion
