+++ Source/Lib/Decoder/Codec/EbDecInterPrediction.c
@@ -34,7 +34,7 @@
 #include "EbDecInterPrediction.h"
 #include "EbUtility.h"
 #include "EbDefinitions.h"
-#include "../../Encoder/Codec/EbWarpedMotion.h"
+#include "EbWarpedMotion.h"
 
 static INLINE void dec_clamp_mv(MV *mv, int32_t min_col, int32_t max_col, int32_t min_row,
                                 int32_t max_row) {
