+++ Source/Lib/Decoder/Codec/EbDecParseInterBlock.c
@@ -16,7 +16,7 @@
 
 #include "EbDecParseInterBlock.h"
 #include "EbCommonUtils.h"
-#include "../../Encoder/Codec/EbWarpedMotion.h"
+#include "EbWarpedMotion.h"
 
 
 typedef const int (*ColorCost)[PALETTE_SIZES][PALETTE_COLOR_INDEX_CONTEXTS][PALETTE_COLORS];
